fix: fix errors, remove unused files
