--- identity/src/deployments/index.ts
@@ -1,13 +1,0 @@
-import _onchainInfo from "./onchain-info.json";
-import _GitcoinAttester from "./abi/GitcoinAttester.json";
-import _GitcoinPassportDecoder from "./abi/GitcoinPassportDecoder.json";
-import _GitcoinResolver from "./abi/GitcoinResolver.json";
-import _GitcoinVeraxPortal from "./abi/GitcoinVeraxPortal.json";
-import _GitcoinVerifier from "./abi/GitcoinVerifier.json";
-
-export const passportOnchainInfo = _onchainInfo;
-export const passportAttesterAbi = _GitcoinAttester;
-export const passportDecoderAbi = _GitcoinPassportDecoder;
-export const passportResolverAbi = _GitcoinResolver;
-export const passportVeraxPortalAbi = _GitcoinVeraxPortal;
-export const passportVerifierAbi = _GitcoinVerifier;
