--- identity/src/deployments/providerBitMapInfo.ts
@@ -1,1 +1,0 @@
-export const msg: string = "Hello, World!";
